LayoutTests/imported/w3c/web-platform-tests/navigation-timing/test_timing_reload-expected.txt

Description

This test validates that the window.performance.timing attributes change when a page is reloaded.

This page should be loaded with a green background frame below. The frame will be automatically reloaded and then verified that the window.performance.timing attributes have been updated to the new reloaded navigation timings.


PASS window.performance is defined
PASS window.performance.navigation is defined
PASS window.performance.navigation.type == TYPE_RELOAD
PASS connectEnd is different after the reload.
PASS connectStart is different after the reload.
PASS domComplete is different after the reload.
PASS domContentLoadedEventEnd is different after the reload.
PASS domContentLoadedEventStart is different after the reload.
PASS domInteractive is different after the reload.
PASS domLoading is different after the reload.
PASS domainLookupEnd is different after the reload.
PASS domainLookupStart is different after the reload.
PASS fetchStart is different after the reload.
PASS loadEventEnd is different after the reload.
PASS loadEventStart is different after the reload.
PASS navigationStart is different after the reload.
PASS requestStart is different after the reload.
PASS responseEnd is different after the reload.
PASS responseStart is different after the reload.
